refactor(hardening): complete Phase 3 structural readability
Problem:
- Large extension, view-provider, GitOps, and webview modules still exceeded the Phase 3 readability target.
- Cognitive-complexity grandfather overrides kept legacy hotspots outside the global gate.

Changed:
- Split activation wiring, commit-context actions, shared panel actions, validation helpers, stash parsing, clone URL parsing, language associations, and undocked webview helpers into focused modules.
- Reduced all source files to 700 lines or fewer.
- Removed per-file cognitive-complexity overrides while keeping the global threshold active.

Preserved:
- Existing command registrations, panel message contracts, Git/stash behavior, clone flow behavior, theme icon resolution, and undocked layout behavior.
- Version remains 0.9.4; no release metadata changes.

Validation:
- bun run format:check
- bun run lint
- bun run lint:strict
- bun run architecture:check
- bun run react-doctor
- bun run typecheck
- bun run build
- bun run l10n:validate && bun run l10n:audit
- bun vitest run focused touched suites
- bun run test
- no src file over 700 lines
- npx gitnexus detect-changes --scope staged
