+++ src/git/operations.ts
@@ -9,8 +9,17 @@ import type {
     MergeConflictFile,
     AmendBranchCommitSummary,
 } from "../types";
-import { getErrorMessage, sanitizeErrorMessage } from "../utils/errors";
+import { getErrorMessage } from "../utils/errors";
 import { isValidBranchName } from "../utils/gitRefs";
+import {
+    assertRepoRelativeGitPath,
+    assertStashIndex,
+    commitStatsUnavailableMessage,
+    getVsCodeApi,
+    logGitOpsWarning,
+    stagedStatsUnavailableMessage,
+    unstagedStatsUnavailableMessage,
+} from "./operationSupport";
 import {
     AMEND_BRANCH_COMMIT_FORMAT,
     COMMIT_DETAIL_FORMAT,
@@ -18,7 +27,6 @@ import {
     isUnmergedConflictCode,
     mapCommitFileStatus,
     mapConflictSideState,
-    mapStatusCode,
     parseAmendBranchCommitSummaries,
     parseCommitDetail,
     parseCommitLog,
@@ -31,122 +39,23 @@ import {
     parseWorkingTreeStatus,
     planRollbackFiles,
 } from "./workingTree";
-
-declare const require: (id: string) => unknown;
-
-const OUTPUT_CHANNEL_NAME = "IntelliGit";
-
-type VsCodeApi = typeof import("vscode");
-type OutputChannelLike = { appendLine: (value: string) => void };
+import { parseShelvedFiles } from "./stashFiles";
 type ConfirmSetUpstreamPush = (remote: string, branch: string) => Promise<boolean>;
-type GitOpsWarningOptions = { userWarningMessage?: string };
-
-let cachedVsCodeApi: VsCodeApi | null | undefined;
-let outputChannel: OutputChannelLike | undefined;
-
-function getVsCodeApi(): VsCodeApi | null {
-    if (cachedVsCodeApi !== undefined) return cachedVsCodeApi;
-    try {
-        cachedVsCodeApi = require("vscode") as VsCodeApi;
-    } catch {
-        cachedVsCodeApi = null;
-    }
-    return cachedVsCodeApi;
-}
-
-function getOutputChannel(): OutputChannelLike {
-    if (outputChannel) return outputChannel;
-    const vscode = getVsCodeApi();
-    outputChannel = vscode
-        ? vscode.window.createOutputChannel(OUTPUT_CHANNEL_NAME)
-        : { appendLine: (value: string) => console.warn(value) };
-    return outputChannel;
-}
-
-function logGitOpsWarning(context: string, err: unknown, options?: GitOpsWarningOptions): void {
-    const channel = getOutputChannel();
-    const message = getErrorMessage(err);
-    channel.appendLine(`[GitOps] ${context}: ${message}`);
-    if (err instanceof Error && err.stack) {
-        channel.appendLine(sanitizeErrorMessage(err.stack));
-    }
-    if (options?.userWarningMessage) {
-        const vscode = getVsCodeApi();
-        if (vscode) {
-            void vscode.window.showWarningMessage(options.userWarningMessage);
-        }
-    }
-}
-
-function commitStatsUnavailableMessage(): string {
-    const vscode = getVsCodeApi();
-    return vscode
-        ? vscode.l10n.t("Some commit change stats may be unavailable.")
-        : "Some commit change stats may be unavailable.";
-}
-
-function unstagedStatsUnavailableMessage(): string {
-    const vscode = getVsCodeApi();
-    return vscode
-        ? vscode.l10n.t("Some unstaged change stats may be unavailable.")
-        : "Some unstaged change stats may be unavailable.";
-}
-
-function stagedStatsUnavailableMessage(): string {
-    const vscode = getVsCodeApi();
-    return vscode
-        ? vscode.l10n.t("Some staged change stats may be unavailable.")
-        : "Some staged change stats may be unavailable.";
-}
-
-function assertStashIndex(index: number): void {
-    if (!Number.isInteger(index) || index < 0) {
-        throw new Error(`Invalid stash index: ${index}`);
-    }
-}
-
-function assertRepoRelativeGitPath(filePath: string): string {
-    const trimmed = filePath.trim();
-    if (
-        !trimmed ||
-        trimmed.startsWith("/") ||
-        trimmed.startsWith("\\") ||
-        /^[a-zA-Z]:[\\/]/.test(trimmed)
-    ) {
-        throw new Error(`Rejected non-relative path: ${filePath}`);
-    }
-    if (/[\0\r\n]/.test(trimmed)) {
-        throw new Error(`Rejected path containing control characters: ${filePath}`);
-    }
-    const normalized = trimmed.replace(/\\/g, "/");
-    const segments = normalized.split("/").filter((segment) => segment && segment !== ".");
-    if (segments.length === 0) {
-        throw new Error(`Rejected repo root path: ${filePath}`);
-    }
-    if (segments.some((segment) => segment === "..")) {
-        throw new Error(`Rejected path escaping repo root: ${filePath}`);
-    }
-    return segments.join("/");
-}
-
 export class UpstreamPushDeclinedError extends Error {
     constructor() {
         super("Upstream push declined by user");
         this.name = "UpstreamPushDeclinedError";
     }
 }
-
 export class GitOps {
     constructor(
         private readonly executor: GitExecutor,
         private readonly confirmSetUpstreamPush?: ConfirmSetUpstreamPush,
     ) {}
-
     async init(repoPath: string): Promise<string> {
         const executor = new GitExecutor(repoPath);
         return executor.run(["init"]);
     }
-
     async isRepository(): Promise<boolean> {
         try {
             await this.executor.run(["rev-parse", "--is-inside-work-tree"]);
@@ -155,7 +64,6 @@ export class GitOps {
             return false;
         }
     }
-
     async hasAnyCommits(): Promise<boolean> {
         try {
             const out = await this.executor.run(["rev-list", "--count", "HEAD"]);
@@ -164,7 +72,6 @@ export class GitOps {
             return false;
         }
     }
-
     async getRemotes(): Promise<string[]> {
         try {
             const out = await this.executor.run(["remote"]);
@@ -177,7 +84,6 @@ export class GitOps {
             return [];
         }
     }
-
     async branchHasUpstream(branch: string): Promise<boolean> {
         try {
             const out = await this.executor.run([
@@ -190,39 +96,30 @@ export class GitOps {
             return false;
         }
     }
-
     async addRemote(name: string, url: string): Promise<void> {
         await this.executor.run(["remote", "add", name, url]);
     }
-
     async removeRemote(name: string): Promise<void> {
         await this.executor.run(["remote", "remove", name]);
     }
-
     async pushWithUpstream(remote: string, branch: string): Promise<string> {
         return this.executor.run(["push", "-u", remote, branch]);
     }
-
     async getRepositoryRoot(): Promise<string> {
         const root = await this.executor.run(["rev-parse", "--show-toplevel"]);
         return root.trim();
     }
-
     async getBranches(): Promise<Branch[]> {
         const format =
             "%(refname)\t%(refname:short)\t%(objectname:short)\t%(upstream:short)\t%(upstream:track,nobracket)\t%(HEAD)";
         const result = await this.executor.run(["branch", "-a", `--format=${format}`]);
-
         const branches: Branch[] = [];
         for (const line of result.trim().split("\n")) {
             if (!line.trim()) continue;
             const [refname, name, hash, upstream, track, head] = line.split("\t");
-
             const isRemote = refname.startsWith("refs/remotes/");
-
             // Skip symbolic refs like origin/HEAD (refname:short resolves to just "origin")
             if (refname.endsWith("/HEAD")) continue;
-
             let remote: string | undefined;
             if (isRemote) {
                 // refname:short for remote is "origin/main", first segment is the remote name
@@ -230,7 +127,6 @@ export class GitOps {
             } else if (upstream) {
                 remote = upstream.split("/")[0];
             }
-
             let ahead = 0,
                 behind = 0;
             if (track) {
@@ -239,7 +135,6 @@ export class GitOps {
                 if (a) ahead = parseInt(a[1]);
                 if (b) behind = parseInt(b[1]);
             }
-
             branches.push({
                 name,
                 hash,
@@ -253,7 +148,6 @@ export class GitOps {
         }
         return branches;
     }
-
     async getLog(
         maxCount: number = 500,
         branch?: string,
@@ -264,13 +158,11 @@ export class GitOps {
         if (skip > 0) {
             args.push(`--skip=${skip}`);
         }
-
         if (filterText) {
             // Use --fixed-strings to treat the filter as a literal string,
             // preventing ReDoS via git's regex engine on user input.
             args.push(`--grep=${filterText}`, "-i", "--fixed-strings");
         }
-
         if (branch) {
             if (!isValidBranchName(branch)) {
                 throw new Error("Invalid branch filter received for git log.");
@@ -279,11 +171,9 @@ export class GitOps {
         } else {
             args.push("--all");
         }
-
         const result = await this.executor.run(args);
         return parseCommitLog(result);
     }
-
     async getUnpushedCommitHashes(): Promise<string[]> {
         try {
             // Commits reachable from local branches but not from any remote-tracking ref.
@@ -298,7 +188,6 @@ export class GitOps {
             return [];
         }
     }
-
     async getCommitDetail(hash: string): Promise<CommitDetail> {
         const info = await this.executor.run([
             "show",
@@ -306,7 +195,6 @@ export class GitOps {
             "--no-patch",
             hash,
         ]);
-
         const filesByPath = new Map<string, CommitFile>();
         const upsertFile = (path: string, status: CommitFile["status"]): CommitFile => {
             const existing = filesByPath.get(path);
@@ -328,7 +216,6 @@ export class GitOps {
             filesByPath.set(path, created);
             return created;
         };
-
         const nameStatus = await this.executor.run([
             "diff-tree",
             "--no-commit-id",
@@ -337,7 +224,6 @@ export class GitOps {
             "--name-status",
             hash,
         ]);
-
         for (const line of nameStatus.trim().split("\n")) {
             if (!line.trim()) continue;
             const cols = line.split("\t");
@@ -349,7 +235,6 @@ export class GitOps {
                 upsertFile(path, status);
             }
         }
-
         try {
             const numstat = await this.executor.run([
                 "diff-tree",
@@ -381,16 +266,12 @@ export class GitOps {
                 userWarningMessage: commitStatsUnavailableMessage(),
             });
         }
-
         return parseCommitDetail(info, hash, Array.from(filesByPath.values()));
     }
-
     // --- Working tree operations ---
-
     async getStatus(): Promise<WorkingFile[]> {
         const result = await this.executor.run(["status", "--porcelain=v1", "-z", "-uall"]);
         const files = parseWorkingTreeStatus(result);
-
         const applyNumstat = (
             output: string,
             staged: boolean,
@@ -403,7 +284,6 @@ export class GitOps {
                 logGitOpsWarning(`Failed to get ${label} numstat`, err, { userWarningMessage });
             }
         };
-
         // Fetch unstaged and staged numstat in parallel
         const [unstagedStat, stagedStat] = await Promise.all([
             this.executor.run(["diff", "--numstat"]).catch((err) => {
@@ -419,69 +299,55 @@ export class GitOps {
                 return "";
             }),
         ]);
-
         applyNumstat(unstagedStat, false, "unstaged", unstagedStatsUnavailableMessage());
         applyNumstat(stagedStat, true, "staged", stagedStatsUnavailableMessage());
-
         return files;
     }
-
     async stageFiles(paths: string[]): Promise<void> {
         if (paths.length === 0) return;
         const pathsToStage = await this.excludeAlreadyStagedDeletedPaths(paths);
         if (pathsToStage.length === 0) return;
         await this.executor.run(["add", "--", ...pathsToStage]);
     }
-
     private async excludeAlreadyStagedDeletedPaths(paths: string[]): Promise<string[]> {
         const status = await this.executor.run(["status", "--porcelain=v1", "-z", "--", ...paths]);
         if (!status.trim()) return paths;
-
         const alreadyStagedDeleted = parseAlreadyStagedDeletedPaths(status);
         return paths.filter((path) => !alreadyStagedDeleted.has(path));
     }
-
     async unstageFiles(paths: string[]): Promise<void> {
         if (paths.length === 0) return;
         await this.executor.run(["reset", "HEAD", "--", ...paths]);
     }
-
     async commit(message: string, amend: boolean = false): Promise<string> {
         const args = ["commit", "-m", message];
         if (amend) args.push("--amend");
         return this.executor.run(args);
     }
-
     async push(): Promise<string> {
         try {
             return await this.executor.run(["push"]);
         } catch (err) {
             if (!isNoUpstreamPushError(err)) throw err;
-
             const suggested = parseSetUpstreamPushSuggestion(err);
             const branch = suggested?.branch ?? (await this.resolveCurrentBranchNameForPush());
             const remote = suggested?.remote ?? (await this.resolveDefaultRemoteNameForPush());
             if (!branch || !remote) throw err;
-
             const allowSetUpstream = await this.requestSetUpstreamPush(remote, branch);
             if (!allowSetUpstream) {
                 throw new UpstreamPushDeclinedError();
             }
-
             return this.executor.run(["push", "--set-upstream", remote, branch]);
         }
     }
-
     async pullRebase(): Promise<string> {
         return this.executor.run(["pull", "--rebase"]);
     }
-
     async commitAndPush(message: string, amend: boolean = false): Promise<string> {
         await this.assertPushRemoteReachable();
         await this.commit(message, amend);
         return this.push();
     }
-
     private async assertPushRemoteReachable(): Promise<void> {
         let upstream: string;
         try {
@@ -491,12 +357,9 @@ export class GitOps {
         } catch {
             return;
         }
-
         if (!upstream || !upstream.includes("/")) return;
-
         const remote = upstream.split("/")[0];
         if (!remote) return;
-
         try {
             await this.executor.run(["ls-remote", "--exit-code", remote]);
         } catch (err) {
@@ -506,7 +369,6 @@ export class GitOps {
             );
         }
     }
-
     private async resolveCurrentBranchNameForPush(): Promise<string | null> {
         try {
             const raw = await this.executor.run(["rev-parse", "--abbrev-ref", "HEAD"]);
@@ -517,7 +379,6 @@ export class GitOps {
             return null;
         }
     }
-
     private async resolveDefaultRemoteNameForPush(): Promise<string | null> {
         try {
             const remotes = await this.executor.run(["remote"]);
@@ -530,15 +391,12 @@ export class GitOps {
             return null;
         }
     }
-
     private async requestSetUpstreamPush(remote: string, branch: string): Promise<boolean> {
         if (this.confirmSetUpstreamPush) {
             return this.confirmSetUpstreamPush(remote, branch);
         }
-
         const vscode = getVsCodeApi();
         if (!vscode) return false;
-
         const confirmLabel = vscode.l10n.t("Set Upstream and Push");
         const selection = await vscode.window.showWarningMessage(
             vscode.l10n.t(
@@ -550,7 +408,6 @@ export class GitOps {
         );
         return selection === confirmLabel;
     }
-
     async getLastCommitMessage(): Promise<string> {
         try {
             return (await this.executor.run(["log", "-1", "--format=%B"])).trim();
@@ -558,7 +415,6 @@ export class GitOps {
             return "";
         }
     }
-
     /**
      * Commits on the current branch relevant when amending: ahead of @{upstream}
      * if set, otherwise the recent history on HEAD (same idea as IntelliJ amend context).
@@ -590,7 +446,6 @@ export class GitOps {
         } catch {
             // No upstream or ambiguous ref — fall through to local history.
         }
-
         try {
             const out = await this.executor.run([
                 "log",
@@ -603,12 +458,10 @@ export class GitOps {
             return [];
         }
     }
-
     async rollbackFiles(paths: string[]): Promise<void> {
         if (paths.length === 0) return;
         const status = await this.executor.run(["status", "--porcelain=v1", "-z", "-uall"]);
         const { resetPaths, checkoutPaths, cleanupPaths } = planRollbackFiles(paths, status);
-
         if (resetPaths.length > 0) {
             await this.executor.run(["reset", "HEAD", "--", ...resetPaths]);
         }
@@ -619,15 +472,12 @@ export class GitOps {
             await this.executor.run(["clean", "-fd", "--", ...cleanupPaths]);
         }
     }
-
     async rollbackAll(): Promise<void> {
         await this.executor.run(["reset", "--hard", "HEAD"]);
         // Also clean untracked files
         await this.executor.run(["clean", "-fd"]);
     }
-
     // --- Shelf operations (implemented via git stash) ---
-
     async shelveSave(paths?: string[], message: string = "Shelved changes"): Promise<string> {
         const args = ["stash", "push", "--include-untracked", "-m", message];
         if (paths && paths.length > 0) {
@@ -635,17 +485,14 @@ export class GitOps {
         }
         return this.executor.run(args);
     }
-
     async shelvePop(index: number = 0): Promise<string> {
         assertStashIndex(index);
         return this.executor.run(["stash", "pop", `stash@{${index}}`]);
     }
-
     async shelveApply(index: number = 0): Promise<string> {
         assertStashIndex(index);
         return this.executor.run(["stash", "apply", `stash@{${index}}`]);
     }
-
     async listShelved(): Promise<StashEntry[]> {
         try {
             const result = await this.executor.run(["stash", "list", "--format=%H\t%gd\t%gs\t%aI"]);
@@ -654,77 +501,32 @@ export class GitOps {
             return [];
         }
     }
-
     async shelveDelete(index: number): Promise<string> {
         assertStashIndex(index);
         return this.executor.run(["stash", "drop", `stash@{${index}}`]);
     }
-
     async getShelvedFiles(index: number): Promise<WorkingFile[]> {
         assertStashIndex(index);
         const ref = `stash@{${index}}`;
-        const files = new Map<string, WorkingFile>();
-
-        const upsert = (path: string, status: WorkingFile["status"] = "M"): WorkingFile => {
-            const existing = files.get(path);
-            if (existing) return existing;
-            const created: WorkingFile = {
-                path,
-                status,
-                staged: false,
-                additions: 0,
-                deletions: 0,
-            };
-            files.set(path, created);
-            return created;
-        };
-
+        let nameStatus = "";
+        let numstat = "";
         try {
-            const nameStatus = await this.executor.run(["stash", "show", "--name-status", ref]);
-            for (const line of nameStatus.trim().split("\n")) {
-                if (!line.trim()) continue;
-                const parts = line.split("\t");
-                if (parts.length < 2) continue;
-                const code = parts[0].trim();
-                const status = mapStatusCode(code[0]) ?? "M";
-                const path =
-                    code.startsWith("R") || code.startsWith("C")
-                        ? (parts[2]?.trim() ?? parts[1]?.trim())
-                        : parts[1]?.trim();
-                if (!path) continue;
-                upsert(path, status);
-            }
+            nameStatus = await this.executor.run(["stash", "show", "--name-status", ref]);
         } catch (err) {
             logGitOpsWarning(`Failed stash show --name-status for ${ref}`, err);
         }
-
         try {
-            const numstat = await this.executor.run(["stash", "show", "--numstat", ref]);
-            for (const line of numstat.trim().split("\n")) {
-                if (!line.trim()) continue;
-                const parts = line.split("\t");
-                if (parts.length < 3) continue;
-                const adds = parts[0] === "-" ? 0 : Number(parts[0]) || 0;
-                const dels = parts[1] === "-" ? 0 : Number(parts[1]) || 0;
-                const path = parts[2].trim();
-                if (!path) continue;
-                const entry = upsert(path);
-                const updated = { ...entry, additions: adds, deletions: dels };
-                files.set(path, updated);
-            }
+            numstat = await this.executor.run(["stash", "show", "--numstat", ref]);
         } catch (err) {
             logGitOpsWarning(`Failed stash show --numstat for ${ref}`, err);
         }
-
-        return Array.from(files.values()).sort((a, b) => a.path.localeCompare(b.path));
+        return parseShelvedFiles(nameStatus, numstat);
     }
-
     async getShelvedFilePatch(index: number, filePath: string): Promise<string> {
         assertStashIndex(index);
         const ref = `stash@{${index}}`;
         return this.executor.run(["diff", `${ref}^`, ref, "--", filePath]);
     }
-
     async getFileHistory(filePath: string, maxCount: number = 50): Promise<string> {
         return this.executor.run([
             "log",
@@ -735,7 +537,6 @@ export class GitOps {
             filePath,
         ]);
     }
-
     async getFileHistoryEntries(
         filePath: string,
         maxCount: number = 30,
@@ -750,10 +551,8 @@ export class GitOps {
             "--",
             filePath,
         ]);
-
         return parseFileHistoryEntries(raw);
     }
-
     async getFileContentAtRef(filePath: string, ref: string): Promise<string> {
         const trimmedRef = ref.trim();
         const safeFilePath = assertRepoRelativeGitPath(filePath);
@@ -766,7 +565,6 @@ export class GitOps {
         }
         return this.executor.run(["show", `${trimmedRef}:${safeFilePath}`]);
     }
-
     async getConflictedFiles(): Promise<string[]> {
         const out = await this.executor.run(["diff", "--name-only", "--diff-filter=U"]);
         return out
@@ -774,7 +572,6 @@ export class GitOps {
             .map((line) => line.trim())
             .filter(Boolean);
     }
-
     async getConflictFilesDetailed(): Promise<MergeConflictFile[]> {
         const result = await this.executor.run(["status", "--porcelain=v1", "-z", "-uall"]);
         const files: MergeConflictFile[] = [];
@@ -782,19 +579,16 @@ export class GitOps {
         for (let i = 0; i < entries.length; i++) {
             const entry = entries[i];
             if (!entry || entry.length < 4) continue;
-
             const oursCode = entry.charAt(0);
             const theirsCode = entry.charAt(1);
             const code = `${oursCode}${theirsCode}`;
             const path = entry.slice(3);
             if (!path) continue;
-
             const isRenameOrCopy =
                 oursCode === "R" || oursCode === "C" || theirsCode === "R" || theirsCode === "C";
             if (isRenameOrCopy && i + 1 < entries.length) {
                 i += 1;
             }
-
             if (!isUnmergedConflictCode(code)) continue;
             files.push({
                 path,
@@ -803,10 +597,8 @@ export class GitOps {
                 theirs: mapConflictSideState(theirsCode),
             });
         }
-
         return files.sort((a, b) => a.path.localeCompare(b.path));
     }
-
     async getConflictFileVersions(
         filePath: string,
     ): Promise<{ base: string; ours: string; theirs: string }> {
@@ -821,7 +613,6 @@ export class GitOps {
                 ),
             ]);
         };
-
         const [base, ours, theirs] = await Promise.all([
             withTimeout(this.executor.run(["show", `:1:${filePath}`]), "base").catch(() => ""),
             withTimeout(this.executor.run(["show", `:2:${filePath}`]), "ours").catch(() => ""),
@@ -829,28 +620,23 @@ export class GitOps {
         ]);
         return { base, ours, theirs };
     }
-
     async stageFile(filePath: string): Promise<void> {
         await this.executor.run(["add", "--", filePath]);
     }
-
     async acceptConflictSide(filePath: string, side: "ours" | "theirs"): Promise<void> {
         const sideArg = side === "ours" ? "--ours" : "--theirs";
         await this.executor.run(["checkout", sideArg, "--", filePath]);
         await this.executor.run(["add", "--", filePath]);
     }
-
     async deleteFile(filePath: string, force: boolean = false): Promise<void> {
         const args = force ? ["rm", "-f", "--", filePath] : ["rm", "--", filePath];
         await this.executor.run(args);
     }
 }
-
 function isNoUpstreamPushError(err: unknown): boolean {
     const message = getErrorMessage(err).toLowerCase();
     return message.includes("has no upstream branch");
 }
-
 function parseSetUpstreamPushSuggestion(err: unknown): { remote: string; branch: string } | null {
     const message = getErrorMessage(err);
     const match = message.match(/git push\s+(?:--set-upstream(?:\s*=\s*|\s+)|-u\s+)(\S+)\s+(\S+)/);
