+++ src/services/cloneService.ts
@@ -5,6 +5,7 @@ import * as fs from "fs/promises";
 import { GitExecutor } from "../git/executor";
 import { getErrorMessage } from "../utils/errors";
 import { runGitCommandWithAskpass } from "./gitAskpass";
+import { extractRepoName } from "./cloneUrl";
 
 // ---------------------------------------------------------------------------
 // Types
@@ -687,33 +688,6 @@ function showCloneErrorMessage(title: string, err: unknown, hints: string[]): vo
     vscode.window.showErrorMessage(
         vscode.l10n.t("{title}: {detail}{hintText}", { title, detail, hintText }),
     );
-}
-
-// ---------------------------------------------------------------------------
-// Helpers
-// ---------------------------------------------------------------------------
-
-function extractRepoName(cloneUrl: string): string {
-    const cleaned = cloneUrl.replace(/\.git$/, "").replace(/\/$/, "");
-    const match = cleaned.match(/\/([^/]+)$/);
-    if (match) return sanitizeRepoDirectoryName(match[1]);
-    const segments = cleaned.split(/[:/]/);
-    return sanitizeRepoDirectoryName(segments[segments.length - 1]);
-}
-
-function sanitizeRepoDirectoryName(value: string | undefined): string {
-    const sanitized = (value ?? "")
-        .replace(/[\\/]/g, "")
-        .split("")
-        .filter((char) => {
-            const code = char.charCodeAt(0);
-            return code >= 32 && code !== 127;
-        })
-        .join("")
-        .replace(/[^a-zA-Z0-9._-]/g, "")
-        .trim();
-    if (!sanitized || sanitized === "." || sanitized === "..") return "repo";
-    return sanitized;
 }
 
 async function runGitCloneWithAskpass(
