+++ src/views/UndockedViewProvider.ts
@@ -1,28 +1,48 @@
 // Manages a vscode.WebviewPanel (editor tab) that combines the commit graph
 // and commit panel into a single unified view. Used when the user enables
 // intelligit.undockableWindow to allow dragging to a second monitor.
-
 import * as vscode from "vscode";
 import { GitOps } from "../git/operations";
 import { IconThemeService } from "./shared";
 import { registerThemeChangeListeners, disposeAll } from "./shared/themeListeners";
 import { buildWebviewShellHtml } from "./webviewHtml";
 import { getErrorMessage } from "../utils/errors";
-import { assertRepoRelativePath, deleteFileWithFallback } from "../utils/fileOps";
-import { runWithNotificationProgress } from "../utils/notifications";
-import { promptRebaseAfterPushRejection, isValidGitHash } from "../services/gitHelpers";
+import { assertRepoRelativePath } from "../utils/fileOps";
 import { isBranchAction, isCommitAction } from "../webviews/protocol/commitGraphTypes";
 import type { Branch, CommitDetail, ThemeFolderIconMap, WorkingFile } from "../types";
 import type { BranchAction, CommitAction } from "../webviews/protocol/commitGraphTypes";
 import type { UnifiedOutbound, UnifiedInbound } from "../webviews/protocol/undockedMessages";
-
+import {
+    assertGitHash,
+    assertNullableString,
+    assertNumber,
+    assertRepoPathArray,
+    assertString,
+} from "./messageValidation";
+import {
+    commitAndPushFromPanel,
+    commitOnlyFromPanel,
+    commitSelectedFromPanel,
+    rollbackFromPanel,
+    shelfMutationFromPanel,
+    shelveSaveFromPanel,
+} from "./commitPanelActions";
+import {
+    deleteFileFromPanel,
+    openFileFromPanel,
+    publishBranchFromPanel,
+    showDiffFromPanel,
+    showHistoryFromPanel,
+    showShelfDiffFromPanel,
+    stageFilesFromPanel,
+    unstageFilesFromPanel,
+} from "./panelFileActions";
 interface PersistedColumnWidths {
     branchWidth: number;
     graphWidth: number;
     infoWidth: number;
     commitPanelWidth: number;
 }
-
 function migratePersistedColumnWidths(value: unknown): PersistedColumnWidths | undefined {
     if (!value || typeof value !== "object") return undefined;
     const saved = value as Record<string, unknown>;
@@ -30,7 +50,6 @@ function migratePersistedColumnWidths(value: unknown): PersistedColumnWidths | u
     const graphWidth = saved.graphWidth;
     const infoWidth = saved.infoWidth;
     const commitPanelWidth = saved.commitPanelWidth;
-
     if (
         typeof branchWidth !== "number" ||
         !Number.isFinite(branchWidth) ||
@@ -41,7 +60,6 @@ function migratePersistedColumnWidths(value: unknown): PersistedColumnWidths | u
     ) {
         return undefined;
     }
-
     return {
         branchWidth,
         graphWidth:
@@ -50,16 +68,13 @@ function migratePersistedColumnWidths(value: unknown): PersistedColumnWidths | u
         commitPanelWidth,
     };
 }
-
 export class UndockedViewProvider {
     public static readonly viewType = "intelligit.undocked";
-
     private panel?: vscode.WebviewPanel;
     private readonly gitOps: GitOps;
     private readonly iconTheme: IconThemeService;
     private repoRootUri: vscode.Uri;
     private repositoryLabel = "";
-
     // Graph-side state
     private currentBranch: string | null = null;
     private filterText = "";
@@ -73,48 +88,37 @@ export class UndockedViewProvider {
     private branchFolderIconsByName: ThemeFolderIconMap = {};
     private commitDetailSeq = 0;
     private themeChangeDisposables: vscode.Disposable[] = [];
-
     // Commit-panel state
     private files: WorkingFile[] = [];
     private lastFileCount = 0;
-
     // Event emitters
     private readonly _onCommitSelected = new vscode.EventEmitter<string>();
     readonly onCommitSelected = this._onCommitSelected.event;
-
     private readonly _onBranchAction = new vscode.EventEmitter<{
         action: BranchAction;
         branchName: string;
     }>();
     readonly onBranchAction = this._onBranchAction.event;
-
     private readonly _onCommitAction = new vscode.EventEmitter<{
         action: CommitAction;
         hash: string;
     }>();
     readonly onCommitAction = this._onCommitAction.event;
-
     private readonly _onOpenCommitFileDiff = new vscode.EventEmitter<{
         commitHash: string;
         filePath: string;
     }>();
     readonly onOpenCommitFileDiff = this._onOpenCommitFileDiff.event;
-
     private readonly _onDidChangeFileCount = new vscode.EventEmitter<number>();
     readonly onDidChangeFileCount = this._onDidChangeFileCount.event;
-
     private readonly _onDidChangeWorkingTree = new vscode.EventEmitter<void>();
     readonly onDidChangeWorkingTree = this._onDidChangeWorkingTree.event;
-
     private readonly _onDockRequested = new vscode.EventEmitter<void>();
     readonly onDockRequested = this._onDockRequested.event;
-
     private readonly _onDidDispose = new vscode.EventEmitter<void>();
     readonly onDidDispose = this._onDidDispose.event;
-
     private static readonly COMMIT_DRAFT_KEY_PREFIX = "commitDraft:";
     private static readonly COLUMN_WIDTHS_KEY = "intelligit.undockedColumnWidths";
-
     constructor(
         private readonly extensionUri: vscode.Uri,
         gitOps: GitOps,
@@ -125,12 +129,10 @@ export class UndockedViewProvider {
         this.repoRootUri = repoRootUri;
         this.iconTheme = new IconThemeService(this.extensionUri);
     }
-
     setRepositoryLabel(label: string): void {
         this.repositoryLabel = label;
         if (this.panel) this.panel.title = `IntelliGit — ${label}`;
     }
-
     setRepositoryRootUri(repoRootUri: vscode.Uri): void {
         this.repoRootUri = repoRootUri;
         this.files = [];
@@ -143,7 +145,6 @@ export class UndockedViewProvider {
         this.offset = 0;
         this.loadingMore = false;
     }
-
     setBranches(branches: Branch[]): void {
         this.branches = branches;
         this.sendBranches().catch((err) => {
@@ -153,7 +154,6 @@ export class UndockedViewProvider {
             );
         });
     }
-
     setCommitDetail(detail: CommitDetail): void {
         const requestId = ++this.commitDetailSeq;
         this.selectedCommitDetail = detail;
@@ -167,33 +167,28 @@ export class UndockedViewProvider {
             );
         });
     }
-
     clearCommitDetail(): void {
         this.commitDetailSeq += 1;
         this.selectedCommitDetail = null;
         this.folderIconsByName = {};
         this.postToWebview({ type: "clearCommitDetail" });
     }
-
     async refresh(): Promise<void> {
         await this.iconTheme.initIconThemeData();
         await this.sendBranches();
         await this.loadInitial();
         await this.refreshCommitPanelData(false);
     }
-
     reveal(): void {
         if (this.panel) {
             this.panel.reveal();
         }
     }
-
     open(): void {
         if (this.panel) {
             this.panel.reveal();
             return;
         }
-
         this.panel = vscode.window.createWebviewPanel(
             UndockedViewProvider.viewType,
             `IntelliGit — ${this.repositoryLabel}`,
@@ -204,18 +199,15 @@ export class UndockedViewProvider {
                 localResourceRoots: [vscode.Uri.joinPath(this.extensionUri, "dist")],
             },
         );
-
         this.iconTheme.attachWebview(this.panel.webview);
         this.registerThemeChangeListeners();
         this.panel.webview.html = this.getHtml(this.panel.webview);
-
         this.panel.onDidDispose(() => {
             this.panel = undefined;
             this.iconTheme.dispose();
             this.disposeThemeChangeDisposables();
             this._onDidDispose.fire();
         });
-
         this.panel.webview.onDidReceiveMessage(async (msg: UnifiedOutbound) => {
             try {
                 await this.handleMessage(msg);
@@ -226,7 +218,6 @@ export class UndockedViewProvider {
             }
         });
     }
-
     dispose(): void {
         this.iconTheme.dispose();
         this.disposeThemeChangeDisposables();
@@ -240,10 +231,21 @@ export class UndockedViewProvider {
         this._onDidDispose.dispose();
         this.panel?.dispose();
     }
-
     // --- Message handling --------------------------------------------------
-
     private async handleMessage(msg: UnifiedOutbound): Promise<void> {
+        const actionDeps = {
+            gitOps: this.gitOps,
+            refreshData: () => this.refreshCommitPanelData(),
+            fireWorkingTreeChanged: () => this._onDidChangeWorkingTree.fire(),
+            postCommitted: () => this.postToWebview({ type: "committed" }),
+            maybeOfferPublishBranch: () => Promise.resolve(),
+        };
+        const fileActionDeps = {
+            gitOps: this.gitOps,
+            getWorkspaceRoot: () => this.repoRootUri,
+            refreshData: () => this.refreshCommitPanelData(),
+            fireWorkingTreeChanged: () => this._onDidChangeWorkingTree.fire(),
+        };
         switch (msg.type) {
             // Graph-side
             case "ready":
@@ -261,21 +263,17 @@ export class UndockedViewProvider {
                     message: this.getStoredCommitDraft(),
                 });
                 break;
-
             case "selectCommit":
-                this._onCommitSelected.fire(this.assertGitHash(msg.hash, "hash"));
+                this._onCommitSelected.fire(assertGitHash(msg.hash, "hash"));
                 break;
-
             case "loadMore":
                 await this.loadMore();
                 break;
-
             case "filterText":
-                await this.filterByText(this.assertString(msg.text, "text"));
+                await this.filterByText(assertString(msg.text, "text"));
                 break;
-
             case "filterBranch":
-                this.currentBranch = this.assertNullableString(msg.branch, "branch");
+                this.currentBranch = assertNullableString(msg.branch, "branch");
                 this.filterText = "";
                 this.postToWebview({
                     type: "setSelectedBranch",
@@ -283,279 +281,122 @@ export class UndockedViewProvider {
                 });
                 await this.loadInitial();
                 break;
-
             case "branchAction":
-                if (!isBranchAction(this.assertString(msg.action, "action"))) {
+                if (!isBranchAction(assertString(msg.action, "action"))) {
                     throw new Error("Invalid branch action received from webview.");
                 }
                 this._onBranchAction.fire({
                     action: msg.action,
-                    branchName: this.assertString(msg.branchName, "branchName"),
+                    branchName: assertString(msg.branchName, "branchName"),
                 });
                 break;
-
             case "commitAction":
-                if (!isCommitAction(this.assertString(msg.action, "action"))) {
+                if (!isCommitAction(assertString(msg.action, "action"))) {
                     throw new Error("Invalid commit action received from webview.");
                 }
                 this._onCommitAction.fire({
                     action: msg.action,
-                    hash: this.assertGitHash(msg.hash, "hash"),
+                    hash: assertGitHash(msg.hash, "hash"),
                 });
                 break;
-
             case "openCommitFileDiff":
                 this._onOpenCommitFileDiff.fire({
-                    commitHash: this.assertGitHash(msg.commitHash, "commitHash"),
-                    filePath: assertRepoRelativePath(this.assertString(msg.filePath, "filePath")),
+                    commitHash: assertGitHash(msg.commitHash, "commitHash"),
+                    filePath: assertRepoRelativePath(assertString(msg.filePath, "filePath")),
                 });
                 break;
-
             case "dock":
                 this._onDockRequested.fire();
                 break;
-
             case "columnWidths":
                 await this.workspaceState?.update(UndockedViewProvider.COLUMN_WIDTHS_KEY, {
-                    branchWidth: this.assertNumber(msg.branchWidth, "branchWidth"),
-                    graphWidth: this.assertNumber(msg.graphWidth, "graphWidth"),
-                    infoWidth: this.assertNumber(msg.infoWidth, "infoWidth"),
-                    commitPanelWidth: this.assertNumber(msg.commitPanelWidth, "commitPanelWidth"),
+                    branchWidth: assertNumber(msg.branchWidth, "branchWidth"),
+                    graphWidth: assertNumber(msg.graphWidth, "graphWidth"),
+                    infoWidth: assertNumber(msg.infoWidth, "infoWidth"),
+                    commitPanelWidth: assertNumber(msg.commitPanelWidth, "commitPanelWidth"),
                 });
                 break;
-
             // Commit-panel-side
             case "refresh":
                 await this.refreshCommitPanelData(false);
                 break;
-
             case "saveCommitDraft": {
-                const message = this.assertString(msg.message, "message");
+                const message = assertString(msg.message, "message");
                 await this.workspaceState?.update(
                     this.getCommitDraftStorageKey(),
                     message || undefined,
                 );
                 break;
             }
-
-            case "stageFiles": {
-                const paths = this.assertRepoPathArray(msg.paths, "paths");
-                await this.gitOps.stageFiles(paths);
-                await this.refreshCommitPanelData();
-                this._onDidChangeWorkingTree.fire();
+            case "stageFiles":
+                await stageFilesFromPanel(fileActionDeps, msg.paths);
                 break;
-            }
-
-            case "unstageFiles": {
-                const paths = this.assertRepoPathArray(msg.paths, "paths");
-                await this.gitOps.unstageFiles(paths);
-                await this.refreshCommitPanelData();
-                this._onDidChangeWorkingTree.fire();
+            case "unstageFiles":
+                await unstageFilesFromPanel(fileActionDeps, msg.paths);
                 break;
-            }
-
             case "commitSelected": {
                 const message = (typeof msg.message === "string" ? msg.message : "").trim();
-                const amend = msg.amend === true;
-                const push = msg.push === true;
-                const paths = this.assertRepoPathArray(msg.paths, "paths");
-                if (!message && !amend) {
-                    vscode.window.showWarningMessage(vscode.l10n.t("Enter a commit message."));
-                    return;
-                }
-                if (paths.length === 0 && !amend) {
-                    vscode.window.showWarningMessage(vscode.l10n.t("Select files to commit."));
-                    return;
-                }
-                if (paths.length > 0) {
-                    await this.gitOps.stageFiles(paths);
-                }
-                try {
-                    const progressTitle = push
-                        ? vscode.l10n.t("Committing and pushing...")
-                        : vscode.l10n.t("Committing...");
-                    await runWithNotificationProgress(progressTitle, async () => {
-                        if (push) {
-                            await this.gitOps.commitAndPush(message, amend);
-                        } else {
-                            await this.gitOps.commit(message, amend);
-                        }
-                    });
-                } catch (err) {
-                    if (
-                        push &&
-                        (await promptRebaseAfterPushRejection(err, this.gitOps, async () => {
-                            await this.gitOps.push();
-                        }))
-                    ) {
-                        this.postToWebview({ type: "committed" });
-                        await this.refreshCommitPanelData();
-                        this._onDidChangeWorkingTree.fire();
-                        return;
-                    }
-                    throw err;
-                }
-                const successMessage = push
-                    ? vscode.l10n.t("Committed and pushed successfully.")
-                    : vscode.l10n.t("Committed successfully.");
-                vscode.window.showInformationMessage(successMessage);
-                this.postToWebview({ type: "committed" });
-                await this.refreshCommitPanelData();
-                this._onDidChangeWorkingTree.fire();
+                await commitSelectedFromPanel(actionDeps, {
+                    message,
+                    amend: msg.amend === true,
+                    push: msg.push === true,
+                    paths: assertRepoPathArray(msg.paths, "paths"),
+                });
                 break;
             }
-
             case "commit": {
                 const message = (typeof msg.message === "string" ? msg.message : "").trim();
-                const amend = msg.amend === true;
-                if (!message && !amend) {
-                    vscode.window.showWarningMessage(vscode.l10n.t("Enter a commit message."));
-                    return;
-                }
-                await runWithNotificationProgress(vscode.l10n.t("Committing..."), async () => {
-                    await this.gitOps.commit(message, amend);
-                });
-                vscode.window.showInformationMessage(vscode.l10n.t("Committed successfully."));
-                this.postToWebview({ type: "committed" });
-                await this.refreshCommitPanelData();
-                this._onDidChangeWorkingTree.fire();
+                await commitOnlyFromPanel(actionDeps, message, msg.amend === true);
                 break;
             }
-
             case "commitAndPush": {
                 const message = (typeof msg.message === "string" ? msg.message : "").trim();
-                const amend = msg.amend === true;
-                if (!message && !amend) {
-                    vscode.window.showWarningMessage(vscode.l10n.t("Enter a commit message."));
-                    return;
-                }
-                try {
-                    await runWithNotificationProgress(
-                        vscode.l10n.t("Committing and pushing..."),
-                        async () => {
-                            await this.gitOps.commitAndPush(message, amend);
-                        },
-                    );
-                } catch (err) {
-                    if (
-                        await promptRebaseAfterPushRejection(err, this.gitOps, async () => {
-                            await this.gitOps.push();
-                        })
-                    ) {
-                        this.postToWebview({ type: "committed" });
-                        await this.refreshCommitPanelData();
-                        this._onDidChangeWorkingTree.fire();
-                        return;
-                    }
-                    throw err;
-                }
-                vscode.window.showInformationMessage(
-                    vscode.l10n.t("Committed and pushed successfully."),
-                );
-                this.postToWebview({ type: "committed" });
-                await this.refreshCommitPanelData();
-                this._onDidChangeWorkingTree.fire();
+                await commitAndPushFromPanel(actionDeps, message, msg.amend === true);
                 break;
             }
-
             case "publishBranch":
-                await vscode.commands.executeCommand("intelligit.publishBranch");
-                await this.refreshCommitPanelData();
-                this._onDidChangeWorkingTree.fire();
+                await publishBranchFromPanel(fileActionDeps);
                 break;
-
             case "getLastCommitMessage": {
                 const lastMsg = await this.gitOps.getLastCommitMessage();
                 this.postToWebview({ type: "lastCommitMessage", message: lastMsg });
                 break;
             }
-
             case "getAmendBranchCommits": {
                 const commits = await this.gitOps.getAmendBranchCommits();
                 this.postToWebview({ type: "amendBranchCommits", commits });
                 break;
             }
-
             case "rollback": {
-                const paths = this.assertRepoPathArray(msg.paths, "paths");
-                const rollbackAction = vscode.l10n.t("Rollback");
-                if (paths.length === 0) {
-                    const confirm = await vscode.window.showWarningMessage(
-                        vscode.l10n.t("Rollback all changes?"),
-                        { modal: true },
-                        rollbackAction,
-                    );
-                    if (confirm !== rollbackAction) return;
-                    await this.gitOps.rollbackAll();
-                } else {
-                    const confirm = await vscode.window.showWarningMessage(
-                        vscode.l10n.t("Rollback {count} file(s)?", { count: paths.length }),
-                        { modal: true },
-                        rollbackAction,
-                    );
-                    if (confirm !== rollbackAction) return;
-                    await this.gitOps.rollbackFiles(paths);
-                }
-                vscode.window.showInformationMessage(vscode.l10n.t("Changes rolled back."));
-                await this.refreshCommitPanelData();
-                this._onDidChangeWorkingTree.fire();
+                await rollbackFromPanel(actionDeps, assertRepoPathArray(msg.paths, "paths"));
                 break;
             }
-
-            case "showDiff": {
-                const filePath = assertRepoRelativePath(this.assertString(msg.path, "path"));
-                const uri = vscode.Uri.joinPath(this.repoRootUri, filePath);
-                await vscode.commands.executeCommand("git.openChange", uri);
+            case "showDiff":
+                await showDiffFromPanel(fileActionDeps, msg.path);
                 break;
-            }
-
             case "shelveSave": {
-                const name = typeof msg.name === "string" ? msg.name : "Shelved changes";
-                let paths: string[] | undefined;
-                if (msg.paths !== undefined) {
-                    paths = this.assertRepoPathArray(msg.paths, "paths");
-                }
-                await this.gitOps.shelveSave(paths, name);
-                vscode.window.showInformationMessage(vscode.l10n.t("Changes shelved."));
-                await this.refreshCommitPanelData();
-                this._onDidChangeWorkingTree.fire();
+                await shelveSaveFromPanel(actionDeps, {
+                    name: typeof msg.name === "string" ? msg.name : "Shelved changes",
+                    paths:
+                        msg.paths !== undefined
+                            ? assertRepoPathArray(msg.paths, "paths")
+                            : undefined,
+                });
                 break;
             }
-
-            case "shelfPop": {
-                const index = this.assertNumber(msg.index, "index");
-                await this.gitOps.shelvePop(index);
-                vscode.window.showInformationMessage(vscode.l10n.t("Unshelved changes."));
-                await this.refreshCommitPanelData();
-                this._onDidChangeWorkingTree.fire();
+            case "shelfPop":
+                await shelfMutationFromPanel(actionDeps, "pop", assertNumber(msg.index, "index"));
                 break;
-            }
-
-            case "shelfApply": {
-                const index = this.assertNumber(msg.index, "index");
-                await this.gitOps.shelveApply(index);
-                vscode.window.showInformationMessage(vscode.l10n.t("Applied shelved changes."));
-                await this.refreshCommitPanelData();
-                this._onDidChangeWorkingTree.fire();
+            case "shelfApply":
+                await shelfMutationFromPanel(actionDeps, "apply", assertNumber(msg.index, "index"));
                 break;
-            }
-
-            case "shelfDelete": {
-                const index = this.assertNumber(msg.index, "index");
-                const deleteAction = vscode.l10n.t("Delete");
-                const confirm = await vscode.window.showWarningMessage(
-                    vscode.l10n.t("Delete this shelved change?"),
-                    { modal: true },
-                    deleteAction,
+            case "shelfDelete":
+                await shelfMutationFromPanel(
+                    actionDeps,
+                    "delete",
+                    assertNumber(msg.index, "index"),
                 );
-                if (confirm !== deleteAction) return;
-                await this.gitOps.shelveDelete(index);
-                vscode.window.showInformationMessage(vscode.l10n.t("Shelved change deleted."));
-                await this.refreshCommitPanelData();
-                this._onDidChangeWorkingTree.fire();
                 break;
-            }
-
             case "shelfSelect": {
                 // The commit-panel side handles shelf selection visually;
                 // we just need to refresh with the new index.
@@ -563,74 +404,29 @@ export class UndockedViewProvider {
                 // fetched when we refresh.
                 break;
             }
-
-            case "showShelfDiff": {
-                const index = this.assertNumber(msg.index, "index");
-                const filePath = assertRepoRelativePath(this.assertString(msg.path, "path"));
-                const patch = await this.gitOps.getShelvedFilePatch(index, filePath);
-                const doc = await vscode.workspace.openTextDocument({
-                    content: patch || `No shelved diff found for ${filePath}.`,
-                    language: "diff",
-                });
-                await vscode.window.showTextDocument(doc, { preview: true });
+            case "showShelfDiff":
+                await showShelfDiffFromPanel(fileActionDeps, msg.index, msg.path);
                 break;
-            }
-
-            case "openFile": {
-                const filePath = assertRepoRelativePath(this.assertString(msg.path, "path"));
-                const uri = vscode.Uri.joinPath(this.repoRootUri, filePath);
-                await vscode.window.showTextDocument(uri);
+            case "openFile":
+                await openFileFromPanel(fileActionDeps, msg.path);
                 break;
-            }
-
-            case "deleteFile": {
-                const filePath = assertRepoRelativePath(this.assertString(msg.path, "path"));
-                const deleteAction = vscode.l10n.t("Delete");
-                const confirm = await vscode.window.showWarningMessage(
-                    vscode.l10n.t("Delete {path}?", { path: filePath }),
-                    { modal: true },
-                    deleteAction,
-                );
-                if (confirm !== deleteAction) return;
-                const deleted = await deleteFileWithFallback(
-                    this.gitOps,
-                    this.repoRootUri,
-                    filePath,
-                );
-                if (!deleted) return;
-                vscode.window.showInformationMessage(
-                    vscode.l10n.t("Deleted {path}", { path: filePath }),
-                );
-                await this.refreshCommitPanelData();
-                this._onDidChangeWorkingTree.fire();
+            case "deleteFile":
+                await deleteFileFromPanel(fileActionDeps, msg.path);
                 break;
-            }
-
-            case "showHistory": {
-                const filePath = assertRepoRelativePath(this.assertString(msg.path, "path"));
-                const history = await this.gitOps.getFileHistory(filePath);
-                const doc = await vscode.workspace.openTextDocument({
-                    content: history || "No history found.",
-                    language: "git-commit",
-                });
-                await vscode.window.showTextDocument(doc, { preview: true });
+            case "showHistory":
+                await showHistoryFromPanel(fileActionDeps, msg.path);
                 break;
-            }
         }
     }
-
     // --- Graph data fetching ------------------------------------------------
-
     private async loadInitial(): Promise<void> {
         const requestId = ++this.requestSeq;
         this.offset = 0;
         this.loadingMore = false;
-
         if (this.currentBranch && !this.branches.some((b) => b.name === this.currentBranch)) {
             this.currentBranch = null;
             this.postToWebview({ type: "setSelectedBranch", branch: null });
         }
-
         try {
             const [commits, unpushedHashes] = await Promise.all([
                 this.gitOps.getLog(
@@ -657,7 +453,6 @@ export class UndockedViewProvider {
             this.postToWebview({ type: "loadError", message });
         }
     }
-
     private async loadMore(): Promise<void> {
         if (this.loadingMore) return;
         this.loadingMore = true;
@@ -692,14 +487,11 @@ export class UndockedViewProvider {
             }
         }
     }
-
     private async filterByText(text: string): Promise<void> {
         this.filterText = text;
         await this.loadInitial();
     }
-
     // --- Commit panel data fetching -----------------------------------------
-
     private async refreshCommitPanelData(silent = false): Promise<void> {
         if (!silent) this.postToWebview({ type: "refreshing", active: true });
         try {
@@ -708,28 +500,23 @@ export class UndockedViewProvider {
             const stashes = await this.gitOps.listShelved();
             const currentBranchHasUpstream = await this.currentBranchHasUpstream();
             const { folderIcons, iconFonts } = this.iconTheme.getThemeData();
-
             // Default to first stash
             const selectedShelfIndex = stashes.length > 0 ? stashes[0].index : null;
-
             const shelfFiles =
                 selectedShelfIndex !== null
                     ? await this.iconTheme.decorateWorkingFiles(
                           await this.gitOps.getShelvedFiles(selectedShelfIndex),
                       )
                     : [];
-
             const cpFolderIconsByName = await this.iconTheme.getFolderIconsByWorkingFiles([
                 ...files,
                 ...shelfFiles,
             ]);
-
             this.files = files;
             const uniquePaths = new Set(files.map((f) => f.path));
             const count = uniquePaths.size;
             this._onDidChangeFileCount.fire(count);
             this.lastFileCount = count;
-
             this.postToWebview({
                 type: "update",
                 files,
@@ -746,9 +533,7 @@ export class UndockedViewProvider {
             if (!silent) this.postToWebview({ type: "refreshing", active: false });
         }
     }
-
     // --- Branch sending -----------------------------------------------------
-
     private async sendBranches(): Promise<void> {
         this.branchFolderIconsByName = await this.iconTheme.getFolderIconsByBranches(this.branches);
         const { folderIcons, iconFonts } = this.iconTheme.getThemeData();
@@ -761,15 +546,12 @@ export class UndockedViewProvider {
             iconFonts,
         });
     }
-
     private async currentBranchHasUpstream(): Promise<boolean> {
         const branches = await this.gitOps.getBranches();
         const currentBranch = branches.find((branch) => branch.isCurrent);
         return currentBranch?.upstream !== undefined && currentBranch.upstream.length > 0;
     }
-
     // --- Commit detail ------------------------------------------------------
-
     private postCommitDetailState(): void {
         const { folderIcons, iconFonts } = this.iconTheme.getThemeData();
         if (this.selectedCommitDetail) {
@@ -785,7 +567,6 @@ export class UndockedViewProvider {
         }
         this.postToWebview({ type: "clearCommitDetail" });
     }
-
     private async decorateAndStoreCommitDetail(
         detail: CommitDetail,
         requestId: number,
@@ -796,9 +577,7 @@ export class UndockedViewProvider {
         this.folderIconsByName = decorated.folderIconsByName;
         this.postCommitDetailState();
     }
-
     // --- HTML generation ----------------------------------------------------
-
     private getHtml(webview: vscode.Webview): string {
         return buildWebviewShellHtml({
             extensionUri: this.extensionUri,
@@ -808,17 +587,13 @@ export class UndockedViewProvider {
             backgroundVar: "var(--vscode-editor-background)",
         });
     }
-
     // --- Commit draft persistence -------------------------------------------
-
     private getCommitDraftStorageKey(): string {
         return `${UndockedViewProvider.COMMIT_DRAFT_KEY_PREFIX}${this.repoRootUri.fsPath}`;
     }
-
     private getStoredCommitDraft(): string {
         return this.workspaceState?.get<string>(this.getCommitDraftStorageKey()) ?? "";
     }
-
     private sendPersistedColumnWidths(): void {
         const saved = migratePersistedColumnWidths(
             this.workspaceState?.get(UndockedViewProvider.COLUMN_WIDTHS_KEY),
@@ -839,23 +614,19 @@ export class UndockedViewProvider {
             });
         }
     }
-
     private sendSettings(): void {
         this.postToWebview({
             type: "settings",
             commitWindowPosition: this.resolveCommitWindowPosition(),
         });
     }
-
     private resolveCommitWindowPosition(): "left" | "right" {
         const config = vscode.workspace.getConfiguration();
         const rawPosition = config.get<string>("intelligit.commitWindowPosition") ?? "auto";
         if (rawPosition === "left" || rawPosition === "right") return rawPosition;
         return config.get<string>("workbench.sideBar.location") === "right" ? "right" : "left";
     }
-
     // --- Theme change listeners ---------------------------------------------
-
     private refreshThemeDataWithErrorHandling(): void {
         this.refreshThemeData().catch((err) => {
             const message = getErrorMessage(err);
@@ -865,7 +636,6 @@ export class UndockedViewProvider {
             this.postToWebview({ type: "error", message });
         });
     }
-
     private async refreshThemeData(): Promise<void> {
         await this.iconTheme.initIconThemeData();
         await this.sendBranches();
@@ -876,7 +646,6 @@ export class UndockedViewProvider {
         const requestId = ++this.commitDetailSeq;
         await this.decorateAndStoreCommitDetail(this.selectedCommitDetail, requestId);
     }
-
     private registerThemeChangeListeners(): void {
         this.themeChangeDisposables.push(
             ...registerThemeChangeListeners(() => this.refreshThemeDataWithErrorHandling()),
@@ -890,58 +659,11 @@ export class UndockedViewProvider {
             }),
         );
     }
-
     private disposeThemeChangeDisposables(): void {
         disposeAll(this.themeChangeDisposables);
     }
-
     // --- Webview helpers ----------------------------------------------------
-
     private postToWebview(msg: UnifiedInbound): void {
         this.panel?.webview.postMessage(msg);
-    }
-
-    // --- Type assertion helpers ---------------------------------------------
-
-    private assertString(value: unknown, field: string): string {
-        if (typeof value !== "string") {
-            throw new Error(`Expected string for '${field}', got ${typeof value}`);
-        }
-        return value;
-    }
-
-    private assertNullableString(value: unknown, field: string): string | null {
-        if (value === null) return null;
-        return this.assertString(value, field);
-    }
-
-    private assertGitHash(value: unknown, field: string): string {
-        const hash = this.assertString(value, field).trim();
-        if (!isValidGitHash(hash)) {
-            throw new Error(`Invalid git hash for '${field}'.`);
-        }
-        return hash;
-    }
-
-    private assertNumber(value: unknown, field: string): number {
-        if (typeof value !== "number" || !Number.isFinite(value)) {
-            throw new Error(`Expected number for '${field}', got ${typeof value}`);
-        }
-        return value;
-    }
-
-    private assertStringArray(value: unknown, field: string): string[] {
-        if (!Array.isArray(value)) {
-            throw new Error(`Expected string[] for '${field}', got ${typeof value}`);
-        }
-        if (!value.every((item): item is string => typeof item === "string")) {
-            throw new Error(`Expected all elements of '${field}' to be strings`);
-        }
-        return value;
-    }
-
-    private assertRepoPathArray(value: unknown, field: string): string[] {
-        const strings = this.assertStringArray(value, field);
-        return strings.map((s) => assertRepoRelativePath(s));
     }
 }
