+++ src/webviews/react/UndockedApp.tsx
@@ -8,15 +8,22 @@ import { ChakraProvider, Box } from "@chakra-ui/react";
 import { BranchColumn } from "./BranchColumn";
 import { CommitList } from "./CommitList";
 import { CommitInfoPane } from "./commit-info/CommitInfoPane";
-import { TabBar } from "./commit-panel/components/TabBar";
-import { CommitTab } from "./commit-panel/components/CommitTab";
-import { ShelfTab } from "./commit-panel/components/ShelfTab";
 import { ThemeIconFontFaces } from "./shared/components";
 import { getVsCodeApi } from "./shared/vscodeApi";
 import { useCheckedFiles } from "./commit-panel/hooks/useCheckedFiles";
 import theme from "./commit-panel/theme";
 import { getSettings } from "./shared/settings";
-import { t } from "./shared/i18n";
+import { CommitPanelPane } from "./undocked/CommitPanelPane";
+import { commitPanelReducer, initialCommitPanelState } from "./undocked/commitPanelState";
+import {
+    computeEqualSectionWidths,
+    migrateSectionWidths,
+    normalizeSectionWidths,
+    sectionWidthsAreClose,
+    type SectionWidths,
+} from "./undocked/sectionWidths";
+import { UndockedHeader } from "./undocked/UndockedHeader";
+import { useColumnPairDrag } from "./undocked/useColumnPairDrag";
 import type {
     Branch,
     Commit,
@@ -24,9 +31,6 @@ import type {
     ThemeFolderIconMap,
     ThemeIconFont,
     ThemeTreeIcon,
-    WorkingFile,
-    StashEntry,
-    AmendBranchCommitSummary,
 } from "../../types";
 import type { BranchAction, CommitAction } from "../protocol/commitGraphTypes";
 import type { UnifiedInbound, UnifiedOutbound } from "../protocol/undockedMessages";
@@ -34,414 +38,6 @@ import type { UnifiedInbound, UnifiedOutbound } from "../protocol/undockedMessag
 // --- Helpers ----------------------------------------------------------------
 
 const vscode = getVsCodeApi<UnifiedOutbound, Record<string, unknown>>();
-
-const MIN_SECTION_WIDTH = 220;
-const DIVIDER_WIDTH = 4;
-const SECTION_COUNT = 4;
-const TOTAL_DIVIDER_WIDTH = 3 * DIVIDER_WIDTH;
-const SECTION_WIDTH_KEYS = ["branchWidth", "graphWidth", "infoWidth", "commitPanelWidth"] as const;
-
-interface SectionWidths {
-    branchWidth: number;
-    graphWidth: number;
-    infoWidth: number;
-    commitPanelWidth: number;
-}
-
-type SectionWidthKey = (typeof SECTION_WIDTH_KEYS)[number];
-
-// Compute equal initial widths for all four sections from the viewport.
-// Sections: Commit | Branches | Graph | Changes (Info)
-// Three dividers (4px each) sit between the four sections.
-function computeEqualSectionWidths(totalWidth?: number): SectionWidths {
-    const fallbackWidth = 300;
-    if (typeof window === "undefined") {
-        return {
-            branchWidth: fallbackWidth,
-            graphWidth: fallbackWidth,
-            infoWidth: fallbackWidth,
-            commitPanelWidth: fallbackWidth,
-        };
-    }
-
-    const available = getAvailableSectionWidth(totalWidth);
-    if (available <= 0) {
-        return {
-            branchWidth: fallbackWidth,
-            graphWidth: fallbackWidth,
-            infoWidth: fallbackWidth,
-            commitPanelWidth: fallbackWidth,
-        };
-    }
-
-    const equalWidth = available / SECTION_COUNT;
-    return {
-        branchWidth: equalWidth,
-        graphWidth: equalWidth,
-        infoWidth: equalWidth,
-        commitPanelWidth: equalWidth,
-    };
-}
-
-function getAvailableSectionWidth(totalWidth?: number): number {
-    if (typeof window === "undefined") return SECTION_COUNT * 300;
-    const containerWidth = typeof totalWidth === "number" ? totalWidth : window.innerWidth;
-    return Math.max(0, containerWidth - TOTAL_DIVIDER_WIDTH);
-}
-
-function sumWidths(widths: SectionWidths): number {
-    return SECTION_WIDTH_KEYS.reduce((total, key) => total + widths[key], 0);
-}
-
-function migrateSectionWidths(value: unknown): SectionWidths | undefined {
-    if (!value || typeof value !== "object") return undefined;
-    const record = value as Record<string, unknown>;
-    const branchWidth = record.branchWidth;
-    const graphWidth = record.graphWidth;
-    const infoWidth = record.infoWidth;
-    const commitPanelWidth = record.commitPanelWidth;
-
-    if (
-        typeof branchWidth !== "number" ||
-        !Number.isFinite(branchWidth) ||
-        typeof infoWidth !== "number" ||
-        !Number.isFinite(infoWidth) ||
-        typeof commitPanelWidth !== "number" ||
-        !Number.isFinite(commitPanelWidth)
-    ) {
-        return undefined;
-    }
-
-    return {
-        branchWidth,
-        graphWidth:
-            typeof graphWidth === "number" && Number.isFinite(graphWidth) ? graphWidth : infoWidth,
-        infoWidth,
-        commitPanelWidth,
-    };
-}
-
-function normalizeSectionWidths(widths: SectionWidths, totalWidth?: number): SectionWidths {
-    const available = getAvailableSectionWidth(totalWidth);
-    if (available <= 0) return computeEqualSectionWidths(totalWidth);
-
-    const rawTotal = sumWidths(widths);
-    if (rawTotal <= 0) return computeEqualSectionWidths(totalWidth);
-
-    const sectionMin = Math.min(MIN_SECTION_WIDTH, available / SECTION_COUNT);
-    let normalized: SectionWidths = {
-        branchWidth: Math.max(sectionMin, widths.branchWidth * (available / rawTotal)),
-        graphWidth: Math.max(sectionMin, widths.graphWidth * (available / rawTotal)),
-        infoWidth: Math.max(sectionMin, widths.infoWidth * (available / rawTotal)),
-        commitPanelWidth: Math.max(sectionMin, widths.commitPanelWidth * (available / rawTotal)),
-    };
-
-    const overflow = sumWidths(normalized) - available;
-    if (overflow <= 0.01) return normalized;
-
-    const reducible = SECTION_WIDTH_KEYS.reduce(
-        (total, key) => total + Math.max(0, normalized[key] - sectionMin),
-        0,
-    );
-    if (reducible <= 0) return computeEqualSectionWidths(totalWidth);
-
-    normalized = SECTION_WIDTH_KEYS.reduce(
-        (next, key) => {
-            const excess = Math.max(0, normalized[key] - sectionMin);
-            next[key] = Math.max(sectionMin, normalized[key] - overflow * (excess / reducible));
-            return next;
-        },
-        { ...normalized },
-    );
-
-    return normalized;
-}
-
-function sectionWidthsAreClose(a: SectionWidths, b: SectionWidths): boolean {
-    return SECTION_WIDTH_KEYS.every((key) => Math.abs(a[key] - b[key]) < 0.5);
-}
-
-function useColumnPairDrag(
-    widths: SectionWidths,
-    setWidths: (widths: SectionWidths) => void,
-    firstKey: SectionWidthKey,
-    secondKey: SectionWidthKey,
-): (e: React.MouseEvent) => void {
-    const draggingRef = useRef(false);
-    const moveRef = useRef<((ev: MouseEvent) => void) | null>(null);
-    const upRef = useRef<(() => void) | null>(null);
-    const widthsRef = useRef(widths);
-    widthsRef.current = widths;
-
-    useEffect(() => {
-        return () => {
-            if (draggingRef.current) {
-                if (moveRef.current) document.removeEventListener("mousemove", moveRef.current);
-                if (upRef.current) document.removeEventListener("mouseup", upRef.current);
-                document.body.style.cursor = "";
-                document.body.style.userSelect = "";
-                draggingRef.current = false;
-            }
-        };
-    }, []);
-
-    return useCallback(
-        (e: React.MouseEvent) => {
-            e.preventDefault();
-            draggingRef.current = true;
-            const startX = e.clientX;
-            const startWidths = widthsRef.current;
-            const firstStart = startWidths[firstKey];
-            const secondStart = startWidths[secondKey];
-            const pairTotal = firstStart + secondStart;
-            const pairMin = Math.min(MIN_SECTION_WIDTH, pairTotal / 2);
-
-            const onMouseMove = (ev: MouseEvent) => {
-                if (!draggingRef.current) return;
-                const delta = ev.clientX - startX;
-                const nextFirst = Math.max(
-                    pairMin,
-                    Math.min(pairTotal - pairMin, firstStart + delta),
-                );
-                setWidths({
-                    ...startWidths,
-                    [firstKey]: nextFirst,
-                    [secondKey]: pairTotal - nextFirst,
-                });
-            };
-
-            const onMouseUp = () => {
-                draggingRef.current = false;
-                moveRef.current = null;
-                upRef.current = null;
-                document.removeEventListener("mousemove", onMouseMove);
-                document.removeEventListener("mouseup", onMouseUp);
-                document.body.style.cursor = "";
-                document.body.style.userSelect = "";
-            };
-
-            moveRef.current = onMouseMove;
-            upRef.current = onMouseUp;
-            document.addEventListener("mousemove", onMouseMove);
-            document.addEventListener("mouseup", onMouseUp);
-            document.body.style.cursor = "col-resize";
-            document.body.style.userSelect = "none";
-        },
-        [firstKey, secondKey, setWidths],
-    );
-}
-
-// --- Commit-panel state reducer (inlined from useExtensionMessages) ----------
-
-interface CommitPanelState {
-    files: WorkingFile[];
-    stashes: StashEntry[];
-    shelfFiles: WorkingFile[];
-    selectedShelfIndex: number | null;
-    folderIcon?: ThemeTreeIcon;
-    folderExpandedIcon?: ThemeTreeIcon;
-    folderIconsByName?: ThemeFolderIconMap;
-    iconFonts: ThemeIconFont[];
-    commitMessage: string;
-    isAmend: boolean;
-    amendBranchCommits: AmendBranchCommitSummary[];
-    amendBranchHistoryLoaded: boolean;
-    isRefreshing: boolean;
-    error: string | null;
-    currentBranchHasUpstream: boolean;
-}
-
-interface CommitPanelPaneProps {
-    width: number;
-    cpState: CommitPanelState;
-    checkedPaths: Set<string>;
-    onToggleFile: (path: string) => void;
-    onToggleFolder: (files: WorkingFile[]) => void;
-    onToggleSection: (files: WorkingFile[]) => void;
-    isAllChecked: (files: WorkingFile[]) => boolean;
-    isSomeChecked: (files: WorkingFile[]) => boolean;
-    onMessageChange: (message: string) => void;
-    onAmendChange: (isAmend: boolean) => void;
-    onCommit: () => void;
-    onCommitAndPush: () => void;
-    currentBranchHasUpstream: boolean;
-    groupByDir: boolean;
-    onToggleGroupBy: () => void;
-}
-
-type CommitPanelAction =
-    | {
-          type: "SET_FILES_AND_STASHES";
-          files: WorkingFile[];
-          stashes: StashEntry[];
-          shelfFiles: WorkingFile[];
-          selectedShelfIndex: number | null;
-          folderIcon?: ThemeTreeIcon;
-          folderExpandedIcon?: ThemeTreeIcon;
-          folderIconsByName?: ThemeFolderIconMap;
-          iconFonts?: ThemeIconFont[];
-          currentBranchHasUpstream: boolean;
-      }
-    | { type: "RESTORE_COMMIT_DRAFT"; message: string }
-    | { type: "SET_LAST_COMMIT_MESSAGE"; message: string }
-    | { type: "COMMITTED" }
-    | { type: "SET_REFRESHING"; active: boolean }
-    | { type: "SET_ERROR"; message: string }
-    | { type: "SET_COMMIT_MESSAGE"; message: string }
-    | { type: "SET_AMEND"; isAmend: boolean }
-    | { type: "SET_AMEND_BRANCH_COMMITS"; commits: AmendBranchCommitSummary[] };
-
-const initialCommitPanelState: CommitPanelState = {
-    files: [],
-    stashes: [],
-    shelfFiles: [],
-    selectedShelfIndex: null,
-    folderIcon: undefined,
-    folderExpandedIcon: undefined,
-    folderIconsByName: undefined,
-    iconFonts: [],
-    commitMessage: "",
-    isAmend: false,
-    amendBranchCommits: [],
-    amendBranchHistoryLoaded: false,
-    isRefreshing: false,
-    error: null,
-    currentBranchHasUpstream: true,
-};
-
-function CommitPanelPane({
-    width,
-    cpState,
-    checkedPaths,
-    onToggleFile,
-    onToggleFolder,
-    onToggleSection,
-    isAllChecked,
-    isSomeChecked,
-    onMessageChange,
-    onAmendChange,
-    onCommit,
-    onCommitAndPush,
-    currentBranchHasUpstream,
-    groupByDir,
-    onToggleGroupBy,
-}: CommitPanelPaneProps): React.ReactElement {
-    return (
-        <Box
-            data-testid="undocked-commit-panel-section"
-            style={{ width: `${width}px` }}
-            flexShrink={0}
-            overflow="hidden"
-            display="flex"
-            flexDirection="column"
-        >
-            <Box flex={1} overflow="hidden" display="flex" flexDirection="column">
-                <TabBar
-                    stashCount={cpState.stashes.length}
-                    commitContent={
-                        <CommitTab
-                            files={cpState.files}
-                            commitMessage={cpState.commitMessage}
-                            isAmend={cpState.isAmend}
-                            amendBranchCommits={cpState.amendBranchCommits}
-                            amendBranchHistoryLoaded={cpState.amendBranchHistoryLoaded}
-                            isRefreshing={cpState.isRefreshing}
-                            checkedPaths={checkedPaths}
-                            onToggleFile={onToggleFile}
-                            onToggleFolder={onToggleFolder}
-                            onToggleSection={onToggleSection}
-                            isAllChecked={isAllChecked}
-                            isSomeChecked={isSomeChecked}
-                            onMessageChange={onMessageChange}
-                            onAmendChange={onAmendChange}
-                            onCommit={onCommit}
-                            onCommitAndPush={onCommitAndPush}
-                            currentBranchHasUpstream={currentBranchHasUpstream}
-                            folderIcon={cpState.folderIcon}
-                            folderExpandedIcon={cpState.folderExpandedIcon}
-                            folderIconsByName={cpState.folderIconsByName}
-                            groupByDir={groupByDir}
-                            onToggleGroupBy={onToggleGroupBy}
-                        />
-                    }
-                    shelfContent={
-                        <ShelfTab
-                            stashes={cpState.stashes}
-                            shelfFiles={cpState.shelfFiles}
-                            selectedIndex={cpState.selectedShelfIndex}
-                            folderIcon={cpState.folderIcon}
-                            folderExpandedIcon={cpState.folderExpandedIcon}
-                            folderIconsByName={cpState.folderIconsByName}
-                            groupByDir={groupByDir}
-                            onToggleGroupBy={onToggleGroupBy}
-                        />
-                    }
-                />
-            </Box>
-        </Box>
-    );
-}
-
-function commitPanelReducer(state: CommitPanelState, action: CommitPanelAction): CommitPanelState {
-    switch (action.type) {
-        case "SET_FILES_AND_STASHES":
-            return {
-                ...state,
-                files: action.files,
-                stashes: action.stashes,
-                shelfFiles: action.shelfFiles,
-                selectedShelfIndex: action.selectedShelfIndex,
-                folderIcon: action.folderIcon ?? state.folderIcon,
-                folderExpandedIcon: action.folderExpandedIcon ?? state.folderExpandedIcon,
-                folderIconsByName: action.folderIconsByName ?? state.folderIconsByName,
-                iconFonts: action.iconFonts ?? state.iconFonts,
-                currentBranchHasUpstream: action.currentBranchHasUpstream,
-                error: null,
-            };
-        case "SET_REFRESHING":
-            if (action.active && state.isAmend) {
-                return {
-                    ...state,
-                    isRefreshing: true,
-                    amendBranchCommits: [],
-                    amendBranchHistoryLoaded: false,
-                };
-            }
-            return { ...state, isRefreshing: action.active };
-        case "RESTORE_COMMIT_DRAFT":
-            return { ...state, commitMessage: action.message };
-        case "SET_LAST_COMMIT_MESSAGE":
-            return { ...state, commitMessage: action.message };
-        case "COMMITTED":
-            return {
-                ...state,
-                commitMessage: "",
-                isAmend: false,
-                amendBranchCommits: [],
-                amendBranchHistoryLoaded: false,
-            };
-        case "SET_ERROR":
-            return { ...state, error: action.message };
-        case "SET_COMMIT_MESSAGE":
-            return { ...state, commitMessage: action.message };
-        case "SET_AMEND":
-            return {
-                ...state,
-                isAmend: action.isAmend,
-                amendBranchCommits: [],
-                amendBranchHistoryLoaded: false,
-            };
-        case "SET_AMEND_BRANCH_COMMITS":
-            if (!state.isAmend) return state;
-            return {
-                ...state,
-                amendBranchCommits: action.commits,
-                amendBranchHistoryLoaded: true,
-            };
-    }
-}
-
-// --- Main App ---------------------------------------------------------------
 
 function App(): React.ReactElement {
     // --- Graph-side state ---
@@ -854,49 +450,7 @@ function App(): React.ReactElement {
         <ChakraProvider theme={theme}>
             <ThemeIconFontFaces fonts={iconFonts} />
             <Box display="flex" height="100vh" overflow="hidden" flexDirection="column">
-                <Box
-                    as="header"
-                    height="32px"
-                    flexShrink={0}
-                    display="flex"
-                    alignItems="center"
-                    justifyContent="space-between"
-                    px="10px"
-                    bg="var(--vscode-sideBar-background)"
-                    borderBottom="1px solid var(--vscode-panel-border)"
-                    color="var(--vscode-foreground)"
-                    fontSize="12px"
-                    fontFamily="var(--vscode-font-family)"
-                >
-                    <Box fontWeight={600}>IntelliGit</Box>
-                    <button
-                        type="button"
-                        onClick={handleDock}
-                        title={t("common.dockIntelliGit")}
-                        aria-label={t("common.dockIntelliGit")}
-                        style={{
-                            display: "inline-flex",
-                            alignItems: "center",
-                            gap: 6,
-                            height: 24,
-                            border: "1px solid var(--vscode-button-border, transparent)",
-                            borderRadius: 3,
-                            padding: "0 8px",
-                            color: "var(--vscode-button-foreground)",
-                            background: "var(--vscode-button-secondaryBackground)",
-                            font: "inherit",
-                            cursor: "pointer",
-                        }}
-                    >
-                        <svg width="14" height="14" viewBox="0 0 16 16" aria-hidden>
-                            <path
-                                fill="currentColor"
-                                d="M2 3h12v10H2V3zm1 1v8h10V4H3zm1 1h3v6H4V5zm4 0h4v2H8V5z"
-                            />
-                        </svg>
-                        {t("common.dock")}
-                    </button>
-                </Box>
+                <UndockedHeader onDock={handleDock} />
                 <Box ref={layoutRef} display="flex" flex={1} overflow="hidden" minHeight={0}>
                     {/* Divider and commit panel — only on left side */}
                     {commitPanelPosition === "left" && (
